requirements.txt
sysrsync
paramiko
gitpython
